Move to a package named more like the core api trace package.
