+++ extensions/incubator/src/main/java/io/opentelemetry/extension/trace/ExtendedTracer.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.extension.tracing;
+package io.opentelemetry.extension.trace;
 
 import io.opentelemetry.api.trace.Span;
 import io.opentelemetry.api.trace.SpanBuilder;
