+++ extensions/incubator/src/test/java/io/opentelemetry/extension/trace/ExtendedTracerTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.extension.tracing;
+package io.opentelemetry.extension.trace;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
